feat: move calculate_intrinsic_gas_after_merge to tx pool (#8914)
