+++ crates/optimism/evm/Cargo.toml
@@ -40,4 +40,5 @@ optimism = [
     "reth-primitives/optimism",
     "reth-provider/optimism",
     "reth-optimism-consensus/optimism",
+    "reth-revm/optimism",
 ]
