+++ crates/revm/Cargo.toml
@@ -36,6 +36,8 @@ tracing.workspace = true
 reth-trie.workspace = true
 
 [features]
-default = ["std"]
+default = ["std", "c-kzg"]
 std = []
+c-kzg = ["revm/c-kzg"]
 test-utils = ["dep:reth-trie"]
+optimism = ["revm/optimism"]
